app_2/requirements.txt
numpy>=1.24
numba>=0.57.0
huggingface-hub==0.24.0
langchain
pydantic
langchain-core
langchain-community
langchain-google-genai
langchain_openai
langchain-huggingface
langchain_experimental
python-dotenv
streamlit
tiktoken
faiss-cpu
mysql-connector-python
pymysql
sentence-transformers
chromadb
google-generativeai
InstructorEmbedding
gradio==4.25.0 
